Last Update before I liberate the Highlands
Computer Vision test code, the images and videos are test videos and images to run the codes through. "Video Code Analyzation Foundation" is the self driving code foundation for video. "Identifying Markers O+Y with Explanation" is code for analyzing distance with still image looking for orange and yellow markers.