Spinnaker still not functional
I've been wrestling with trying to get the camera driver working, the current theory is that the SDK version is current. (2.7) but the ROS driver is only confirmed to work with version 2.2 and after 2.3  there are known compatibility problems. I will need to continue to work with this. Looks like I need to hunt down the correct version