Testing joy_driver2 & roboclaw problems
The joy_driver2 node worked great, however, there is a greater need for button debouncing than originally planned. Roboclaw_driver seems to have an issues with the python library. Gotta sort that out. Latest problem when trying to run examples was that the address variable needed to be encoded into bytes.