Begin BruceGuide & begin multicamAcquisition
Bruce guide is a long time coming. right now its just a copy of BettyGuide but Iwill do a lot of the editing on github

BrobotAcquisition.py seeks to be a better implementation for both cameras. its not functional yet, but when finished it should remove any impedance of the brobot sample rate