Acquisition overhaul (pre-testing)
This hasn't been tested yet but the process flow diagram for it is on page 2 of my lab notebook.

This is based on a faster loop that, instead of waiting for a networked message from ROS to grab the image from buffer, spins as fast as possible, in order to grab as many images off the buffer as possible.